+++ src/Controller/AbstractRestfulController.php
@@ -45,9 +45,23 @@ abstract class AbstractRestfulController extends AbstractController
     protected $identifierName = 'id';
 
     /**
-     * @var int From Zend\Json\Json
+     * Flag to pass to json_decode and/or Zend\Json\Json::decode.
+     *
+     * The flags in Zend\Json\Json::decode are integers, but when evaluated
+     * in a boolean context map to the flag passed as the second parameter
+     * to json_decode(). As such, you can specify either the Zend\Json\Json
+     * constant or the boolean value. By default, starting in v3, we use
+     * the boolean value, and cast to integer if using Zend\Json\Json::decode.
+     *
+     * Default value is boolean true, meaning JSON should be cast to
+     * associative arrays (vs objects).
+     *
+     * Override the value in an extending class to set the default behavior
+     * for your class.
+     *
+     * @var int|bool
      */
-    protected $jsonDecodeType = Json::TYPE_ARRAY;
+    protected $jsonDecodeType = true;
 
     /**
      * Map of custom HTTP methods and their handlers
@@ -293,8 +307,7 @@ abstract class AbstractRestfulController extends AbstractController
     public function dispatch(Request $request, Response $response = null)
     {
         if (! $request instanceof HttpRequest) {
-            throw new Exception\InvalidArgumentException(
-                    'Expected an HTTP request');
+            throw new Exception\InvalidArgumentException('Expected an HTTP request');
         }
 
         return parent::dispatch($request, $response);
@@ -316,8 +329,7 @@ abstract class AbstractRestfulController extends AbstractController
              * @todo Determine requirements for when route match is missing.
              *       Potentially allow pulling directly from request metadata?
              */
-            throw new Exception\DomainException(
-                    'Missing route matches; unsure how to retrieve action');
+            throw new Exception\DomainException('Missing route matches; unsure how to retrieve action');
         }
 
         $request = $e->getRequest();
@@ -446,16 +458,16 @@ abstract class AbstractRestfulController extends AbstractController
      *
      * @param Request $request
      * @return mixed
+     * @throws Exception\DomainException If a JSON request was made, but no
+     *    method for parsing JSON is available.
      */
     public function processPostData(Request $request)
     {
         if ($this->requestHasContentType($request, self::CONTENT_TYPE_JSON)) {
-            $data = Json::decode($request->getContent(), $this->jsonDecodeType);
-        } else {
-            $data = $request->getPost()->toArray();
+            return $this->create($this->jsonDecode($request->getContent()));
         }
 
-        return $this->create($data);
+        return $this->create($request->getPost()->toArray());
     }
 
     /**
@@ -535,7 +547,7 @@ abstract class AbstractRestfulController extends AbstractController
      * Attempts to see if an identifier was passed in either the URI or the
      * query string, returning it if found. Otherwise, returns a boolean false.
      *
-     * @param  \Zend\Mvc\Router\RouteMatch $routeMatch
+     * @param  \Zend\Router\RouteMatch $routeMatch
      * @param  Request $request
      * @return false|mixed
      */
@@ -566,6 +578,8 @@ abstract class AbstractRestfulController extends AbstractController
      *
      * @param  mixed $request
      * @return object|string|array
+     * @throws Exception\DomainException If a JSON request was made, but no
+     *    method for parsing JSON is available.
      */
     protected function processBodyContent($request)
     {
@@ -573,7 +587,7 @@ abstract class AbstractRestfulController extends AbstractController
 
         // JSON content? decode and return it.
         if ($this->requestHasContentType($request, self::CONTENT_TYPE_JSON)) {
-            return Json::decode($content, $this->jsonDecodeType);
+            return $this->jsonDecode($request->getContent());
         }
 
         parse_str($content, $parsedParams);
@@ -586,5 +600,36 @@ abstract class AbstractRestfulController extends AbstractController
         }
 
         return $parsedParams;
+    }
+
+    /**
+     * Decode a JSON string.
+     *
+     * Uses json_decode by default. If that is not available, checks for
+     * availability of Zend\Json\Json, and uses that if present.
+     *
+     * Otherwise, raises an exception.
+     *
+     * Marked protected to allow usage from extending classes.
+     *
+     * @param string
+     * @return mixed
+     * @throws Exception\DomainException if no JSON decoding functionality is
+     *     available.
+     */
+    protected function jsonDecode($string)
+    {
+        if (function_exists('json_decode')) {
+            return json_decode($string, (bool) $this->jsonDecodeType);
+        }
+
+        if (class_exists(Json::class)) {
+            return Json::decode($string, (int) $this->jsonDecodeType);
+        }
+
+        throw new Exception\DomainException(sprintf(
+            'Unable to parse JSON request, due to missing ext/json and/or %s',
+            Json::class
+        ));
     }
 }