+++ src/Service/ResponseFactory.php
@@ -10,42 +10,35 @@
 namespace Zend\Mvc\Service;
 
 use Interop\Container\ContainerInterface;
-use Zend\Console\Console;
-use Zend\Console\Response as ConsoleResponse;
 use Zend\Http\PhpEnvironment\Response as HttpResponse;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
-use Zend\Stdlib\MessageInterface;
 
 class ResponseFactory implements FactoryInterface
 {
     /**
-     * Create and return a response instance, according to current environment.
+     * Create and return a response instance.
      *
      * @param  ContainerInterface $container
      * @param  string $name
      * @param  null|array $options
-     * @return MessageInterface
+     * @return HttpResponse
      */
     public function __invoke(ContainerInterface $container, $name, array $options = null)
     {
-        if (Console::isConsole()) {
-            return new ConsoleResponse();
-        }
-
         return new HttpResponse();
     }
 
     /**
-     * Create and return MessageInterface instance
+     * Create and return response instance.
      *
      * For use with zend-servicemanager v2; proxies to __invoke().
      *
      * @param ServiceLocatorInterface $container
-     * @return MessageInterface
+     * @return HttpResponse
      */
     public function createService(ServiceLocatorInterface $container)
     {
-        return $this($container, MessageInterface::class);
+        return $this($container, 'Response');
     }
 }