+++ test/MiddlewareListenerTest.php
@@ -19,7 +19,7 @@ use Zend\Http\Response;
 use Zend\Mvc\Application;
 use Zend\Mvc\MiddlewareListener;
 use Zend\Mvc\MvcEvent;
-use Zend\Mvc\Router\RouteMatch;
+use Zend\Router\RouteMatch;
 use Zend\ServiceManager\ServiceManager;
 
 class MiddlewareListenerTest extends TestCase