+++ test/Service/ViewHelperManagerFactoryTest.php
@@ -10,15 +10,12 @@
 namespace ZendTest\Mvc\Service;
 
 use PHPUnit_Framework_TestCase as TestCase;
-use ReflectionProperty;
-use Zend\Console\Console;
-use Zend\Console\Request as ConsoleRequest;
 use Zend\Http\PhpEnvironment\Request;
 use Zend\Mvc\Application;
 use Zend\Mvc\MvcEvent;
-use Zend\Mvc\Router\RouteMatch;
-use Zend\Mvc\Router\RouteStackInterface;
 use Zend\Mvc\Service\ViewHelperManagerFactory;
+use Zend\Router\RouteMatch;
+use Zend\Router\RouteStackInterface;
 use Zend\ServiceManager\ServiceManager;
 use Zend\View\Helper;
 
@@ -56,46 +53,6 @@ class ViewHelperManagerFactoryTest extends TestCase
         $this->assertInstanceof('Zend\View\Helper\Doctype', $doctype);
     }
 
-    public function testConsoleRequestsResultInSilentFailure()
-    {
-        $this->services->setService('config', []);
-        $this->services->setService('Request', new ConsoleRequest());
-
-        $manager = $this->factory->__invoke($this->services, 'ViewHelperManager');
-
-        $doctype = $manager->get('doctype');
-        $this->assertInstanceof('Zend\View\Helper\Doctype', $doctype);
-
-        $basePath = $manager->get('basepath');
-        $this->assertInstanceof('Zend\View\Helper\BasePath', $basePath);
-    }
-
-    /**
-     * @group 6247
-     */
-    public function testConsoleRequestWithBasePathConsole()
-    {
-        // Force Console context
-        $r = new ReflectionProperty(Console::class, 'isConsole');
-        $r->setAccessible(true);
-        $r->setValue(true);
-
-        if (! Console::isConsole()) {
-            $this->markTestSkipped('Cannot force console context; skipping test');
-        }
-
-        $this->services->setService('config', [
-            'view_manager' => [
-                'base_path_console' => 'http://test.com'
-            ]
-        ]);
-
-        $manager = $this->factory->__invoke($this->services, 'ViewHelperManager');
-
-        $basePath = $manager->get('basepath');
-        $this->assertEquals('http://test.com', $basePath());
-    }
-
     public function urlHelperNames()
     {
         return [
@@ -112,6 +69,12 @@ class ViewHelperManagerFactoryTest extends TestCase
      */
     public function testUrlHelperFactoryCanBeInvokedViaShortNameOrFullClassName($name)
     {
+        $this->markTestSkipped(sprintf(
+            '%s::%s skipped until zend-view and the url() view helper are updated to use zend-router',
+            get_class($this),
+            __FUNCTION__
+        ));
+
         $routeMatch = $this->prophesize(RouteMatch::class)->reveal();
         $mvcEvent = $this->prophesize(MvcEvent::class);
         $mvcEvent->getRouteMatch()->willReturn($routeMatch);
@@ -138,14 +101,6 @@ class ViewHelperManagerFactoryTest extends TestCase
         $names = ['basepath', 'basePath', 'BasePath', Helper\BasePath::class, 'zendviewhelperbasepath'];
 
         $configurations = [
-            'console' => [[
-                'config' => [
-                    'view_manager' => [
-                        'base_path_console' => '/foo/bar',
-                    ],
-                ],
-            ], '/foo/bar'],
-
             'hard-coded' => [[
                 'config' => [
                     'view_manager' => [